fix bugs #2

+++ FirstTransitionHeader.h
@@ -4,7 +4,7 @@ extern char *instructions[];
 extern char *registerNames[];
 extern int flagLabel;
 typedef struct binWord{
-	short a : 10;
+	unsigned int a : 10;
 }binWord;
 
 void analizeLine(char *line, int lineNum);
